+++ Project/SourceCode/Tab/game_clear_tab.cpp
@@ -49,7 +49,7 @@ GameClearTab::GameClearTab() :
 	for (size_t i = 0; i < data.explanatory_text_data.size(); ++i)
 	{
 		text::CreateText(data.explanatory_text_data.at(i));
-		m_button_prompt->AddExplanatoryText(i, data.explanatory_text_data.at(i).text);
+		m_button_prompt->AddExplanatoryText(static_cast<int>(i), data.explanatory_text_data.at(i).text);
 	}
 
 	CalcAlphaBlendNum();
